requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
PyJWT>=2.8.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.9.0
